allow Coreplex to be on a separate clock
