+++ src/main/scala/rocketchip/TestConfigs.scala
@@ -26,7 +26,8 @@ class WithUnitTest extends Config(
         DefaultTestSuites.groundtest32
       TestGeneration.addSuite(groundtest("p"))
       TestGeneration.addSuite(DefaultTestSuites.emptyBmarks)
-      (p: Parameters, c: CoreplexConfig) => Module(new UnitTestCoreplex(p, c))
+      (p: Parameters, c: CoreplexConfig, clock: Clock, reset: Bool) =>
+        Module(new UnitTestCoreplex(p, c, clock, reset))
     }
     case UnitTests => (testParams: Parameters) =>
       JunctionsUnitTests(testParams) ++ UncoreUnitTests(testParams)
@@ -43,7 +44,8 @@ class UnitTestConfig extends Config(new WithUnitTest ++ new BaseConfig)
 class WithGroundTest extends Config(
   (pname, site, here) => pname match {
     case BuildCoreplex =>
-      (p: Parameters, c: CoreplexConfig) => Module(new GroundTestCoreplex(p, c))
+      (p: Parameters, c: CoreplexConfig, clock: Clock, reset: Bool) =>
+        Module(new GroundTestCoreplex(p, c, clock, reset))
     case TLKey("L1toL2") => {
       val useMEI = site(NTiles) <= 1 && site(NCachedTileLinkPorts) <= 1
       TileLinkParameters(
